+++ simapp/app.go
@@ -22,7 +22,6 @@ import (
 	"github.com/line/lbm-sdk/client"
 	nodeservice "github.com/line/lbm-sdk/client/grpc/node"
 	"github.com/line/lbm-sdk/client/grpc/tmservice"
-	"github.com/line/lbm-sdk/client/rpc"
 	"github.com/line/lbm-sdk/codec"
 	"github.com/line/lbm-sdk/codec/types"
 	"github.com/line/lbm-sdk/server/api"
@@ -36,7 +35,6 @@ import (
 	"github.com/line/lbm-sdk/version"
 	"github.com/line/lbm-sdk/x/auth"
 	"github.com/line/lbm-sdk/x/auth/ante"
-	authrest "github.com/line/lbm-sdk/x/auth/client/rest"
 	authkeeper "github.com/line/lbm-sdk/x/auth/keeper"
 	authsims "github.com/line/lbm-sdk/x/auth/simulation"
 	authtx "github.com/line/lbm-sdk/x/auth/tx"
@@ -86,7 +84,7 @@ import (
 	icahostkeeper "github.com/line/lbm-sdk/x/ibc/applications/27-interchain-accounts/host/keeper"
 	icahosttypes "github.com/line/lbm-sdk/x/ibc/applications/27-interchain-accounts/host/types"
 	icatypes "github.com/line/lbm-sdk/x/ibc/applications/27-interchain-accounts/types"
-	transfer "github.com/line/lbm-sdk/x/ibc/applications/transfer"
+	"github.com/line/lbm-sdk/x/ibc/applications/transfer"
 	ibctransferkeeper "github.com/line/lbm-sdk/x/ibc/applications/transfer/keeper"
 	ibctransfertypes "github.com/line/lbm-sdk/x/ibc/applications/transfer/types"
 	ibc "github.com/line/lbm-sdk/x/ibc/core"
@@ -182,7 +180,6 @@ var (
 		distrtypes.ModuleName:          nil,
 		foundation.ModuleName:          nil,
 		foundation.TreasuryName:        nil,
-		foundation.GovMinterName:       {authtypes.Minter},
 		minttypes.ModuleName:           {authtypes.Minter},
 		stakingtypes.BondedPoolName:    {authtypes.Burner, authtypes.Staking},
 		stakingtypes.NotBondedPoolName: {authtypes.Burner, authtypes.Staking},
@@ -856,9 +853,7 @@ func (app *SimApp) SimulationManager() *module.SimulationManager {
 // API server.
 func (app *SimApp) RegisterAPIRoutes(apiSvr *api.Server, apiConfig config.APIConfig) {
 	clientCtx := apiSvr.ClientCtx
-	rpc.RegisterRoutes(clientCtx, apiSvr.Router)
-	// Register legacy tx routes.
-	authrest.RegisterTxRoutes(clientCtx, apiSvr.Router)
+
 	// Register new tx routes from grpc-gateway.
 	authtx.RegisterGRPCGatewayRoutes(clientCtx, apiSvr.GRPCGatewayRouter)
 	// Register new tendermint queries routes from grpc-gateway.
@@ -867,8 +862,7 @@ func (app *SimApp) RegisterAPIRoutes(apiSvr *api.Server, apiConfig config.APICon
 	// Register node gRPC service for grpc-gateway.
 	nodeservice.RegisterGRPCGatewayRoutes(clientCtx, apiSvr.GRPCGatewayRouter)
 
-	// Register legacy and grpc-gateway routes for all modules.
-	ModuleBasics.RegisterRESTRoutes(clientCtx, apiSvr.Router)
+	// Register grpc-gateway routes for all modules.
 	ModuleBasics.RegisterGRPCGatewayRoutes(clientCtx, apiSvr.GRPCGatewayRouter)
 
 	// register swagger API from root so that other applications can override easily