+++ CHANGELOG.md
@@ -35,16 +35,24 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 # Changelog
 
-## [Unreleased](https://github.com/line/lbm-sdk/compare/v0.46.0-rc10...HEAD)
+## [Unreleased](https://github.com/line/lbm-sdk/compare/v0.46.0...HEAD)
 
 ### Features
+* (x/foundation) [\#848](https://github.com/line/lbm-sdk/pull/848) remove `gov mint` for x/foundation proposal
 
 ### Improvements
+* (cosmovisor) [\#792](https://github.com/line/lbm-sdk/pull/792) Use upstream's cosmovisor
+* (server) [\#821](https://github.com/line/lbm-sdk/pull/821) Get validator pubkey considering KMS
 
 ### Bug Fixes
+* (client) [\#817](https://github.com/line/lbm-sdk/pull/817) remove support for composite (BLS) type
+
+* (x/foundation) [#834](https://github.com/line/lbm-sdk/pull/834) Apply foundation audit
 
 ### Breaking Changes
+* (rest) [\#807](https://github.com/line/lbm-sdk/pull/807) remove legacy REST API
 
 ### Build, CI
+* (ci) [\#829](https://github.com/line/lbm-sdk/pull/829) automate release process
 
 ### Document Updates